requirements.txt
Pyside6
ffmpeg-python==0.2.0
onnxruntime==1.18.1
onnx==1.15.0
psutil
pyinstaller==6.14.0
numpy